+++ stdlib/REPL/test/replcompletions.jl
@@ -7,7 +7,7 @@ using REPL
     @testset "Check symbols previously not shown by REPL.doc_completions()" begin
     symbols = ["?","=","[]","[","]","{}","{","}",";","","'","&&","||","julia","Julia","new","@var_str"]
         for i in symbols
-            @test REPL.doc_completions(i)[1]==i
+            @test REPL.doc_completions(i, Main)[1]==i
         end
     end
 let ex = quote